requirements.txt
streamlit==1.46.1
groq==0.30.0
python-dotenv==1.0.1
